Admin View (#7)
If you go to "/", it will automatically redirect you to "/products" or "/admin" depending on your role. We will later add the appropriate nav links to show/hide accordingly. 

If you don't have access to "/admin" and you go to that route directly, it will not redirect you, but will show you a "cannot display route" message.